Created an install_mysql.pp script. This uses placeholder variables for all library names, install directories and fileserver paths. This will install the newer mysql shared compatibility libary followed by uninstalling the older shared libaries that are depended on by other components.
The client, shared and server mysql libraries are then installed.

A selection of Exec commands are used to stop the mysql service and
restart it in safe mode with the --skip-grant-tables switch. This allows
for us to set the root password to our $password variable.
The service is then stopped to exit safe mode and then restarted.