Created an install_tomcat.pp, this manages the importing of setup files to the target machine, creation of the tomcat users and the starting up of tomcat.
Later on we should specify that tomcat start up as a service.
tomcat-users.xml is included in the /installers folder to setup two
default users; manager and robomanager for handling the gui and the
scripted access respectively. This will be templated in future to allow
for passing of new passwords and usernames.

Unpack tomcat.sh is used to handle the unzipping, unpacking and copying
of the tomcat7 folder structure to /var/hosting/tomcat7. It also handles
cleanup of the residual .tar file and intermediatory directories.