Now disable iptables service so we can access tomcat's manager page.
Setup vagrant file port forwarding from host 8081 to 8080 for the tomcat
manager page.

Added a tomcat-users.xml.erb template to create a manager user with
customisable password, currently hardcoded in the module.

Tomcat is started from the CATALINA_HOME/bin/startup.sh. Service startup
to follow.

Java is now specified as a required module at the beginning of the
tomcat module.

local_install_dir is now a parameter with a suitable default, this is to
prevent clashes with the same java declaration.