Created an iptables module.
Currently it will specify a rule for a port number on either udp or tcp
protocols.

This rule will be *inserted* into the top of the rules list, bumping
everything down.

Then the iptables in-memory store is flushed to /etc/sysconfig/iptables.

The service is then restarted.

Multiple executions of this class with the same args will result in
duplicate rules.