Updated iptables to CentOS 7.
Added a vagrant profile "centos7" for testing.

Added testing procedure to the README.markdown file.

Included bash scripts for performing the iptables config and rollback as
examples.

Added a conditional section to the main iptables module for the case
where we have centos7, this mostly managed install iptables as a service
and disabling firewalld.