Added unless and onlyif clauses to alternatives::install and alternatives::set respectively.
Alternatves::install should only run if grepping for the location of the
jdk within the output of the alternatives --display fails to return
true, in other words if the jdk already exists then we don't do
anything.

Alternatives::set should only run if grepping for the executable name
within the output of alternatives --display returns a value, in other
word it only runs when there is an alternative for the executable
installed.