Initial commit of Apache2 Ubuntu installation on Ubuntu 15.10
Had to remove dependency between http class and the iptables class as it
isn't enabled by default on Ubuntu, instead it uses ufw.

Will need to either modify the dependency between modules to include a
ufw module for ubuntu or install iptables as part of httpd, I prefer the
former approach.