Updated x-clacks module to clarify a requirement on the header::install definition.
Updated the CSP for virtual hosts to fix the requirements on restarting apache.

Reclarified the IfModule 'mod_headers.c'/Directory insertion to add a new directory after the last  directory enabling multiple directories to be added.

Updated the virtual_host tests for CSP with the addition of the x-clacks header.